experiments/cocycle/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cocycle-experiments"
version = "0.1.0"
description = "MPS harness for 'Cocycle Interactions' NeurIPS experiments"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "transformer_lens>=2.2.2",
    "transformers>=4.38.0",
    "sae_lens>=3.0.0",
    "datasets>=2.18.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["harness*", "scripts*"]
